Fixed multiline Inside Pair
The shrink function would fail on multiline targets. It seems it was not
changed after multiline support was introduced.

Shrink only searched wackwards and forwards for whitespace in the
current line. The searchpos call has been updated to use the start and
end line of the target. The edge case where there is only or no
whitespace has been adjusted as well.