Fixed Bug when selecting Inside Tag
The 'It' text object uses shrink to drop whitespace. However, shrink
expects the delimiters to exist, which are dropped by dropt. This causes
an additional character to be dropped left and right.
vIt selects bcd in <a>abcde</a> instead of abcde. The solution is to not
drop the delimiters in dropt so that they can be dropped by shrink
instead. As this breaks 'it', we have to call drop after dropt to drop
the delimiters there.