Fixed removal of nlNL targets to break mappings
When any of the nlNL targets were disabled, ie. set to ' ', some
standard operators like 'a' could break as a result. Instead of creating
a mapping like 'aL', only the operator mapping containing only 'a' would be created,
overwriting the default. This fix only applies nlNL mappings if they
were not disabled by the user.