Improve and simplify quote text objects
Instead of halfway scanning the cursor line for quotation characters, we
now just count how many there are before, after and on the cursor to
decide at what rate to iterate quote targets and whether to skip
initially per direction.

Don't try to be smart if there are no quotation characters in the
cursorline or the total number is odd.