Fix substring extraction in old Vim
Apparently Vim versions 7.3 and 7.4 don't manage to properly parse a[b:]
where a is a list or string and b is an index, while Vim 8 and Neovim
handle it just fine. Interestingly a[2:] works in all versions.

This commit adds a space to not confuse old Vim versions, and use a[b :]
instead.