Remove some outdated, weirdly behaving tests
These tests used to test unsupported cases, and these cases are still
unsupported and work as expected in practice (nothing happens if you
type d2a, for example). But for some reason when executed in the tests
they show unexpected results.

I reduced it to a smaller vim script which shows the issue, it seems
related to cancelling operator pending mode mappings when invoked as
`vim -S script`. But it seems so niche that I'm not planning to report
this. Let me know if you would want me to.