Fix test for whether GenFunc is a function
The current test for whether GenFunc is a function is:

    if GenFunc == 0

On up-to-date versions of vim, this works fine. However, on older
versions (in particular vim 7.4 with patches 0-900) this throws an
"Invalid expression" error. Rather, targets should test directly whether
GenFunc is a FunctionRef. This patch does that as recommended in `:h
type()`.